+++ eframe/CHANGELOG.md
@@ -10,7 +10,7 @@ NOTE: [`egui-winit`](../egui-winit/CHANGELOG.md), [`egui_glium`](../egui_glium/C
   * Add features "wgpu" and "glow"
   * Add `NativeOptions::renderer` to switch between the rendering backends
 * Fix clipboard on Wayland ([#1613](https://github.com/emilk/egui/pull/1613)).
-
+* Allow running on native without hardware accelerated rendering. Change with `NativeOptions::hardware_acceleration` ([#1681]([#1693](https://github.com/emilk/egui/pull/1693)).
 
 ## 0.18.0 - 2022-04-30
 * MSRV (Minimum Supported Rust Version) is now `1.60.0` ([#1467](https://github.com/emilk/egui/pull/1467)).